add build directive to honor noasm build tag
This commit adds a build directive to honor the
`noasm` build tag and fixes the build tag directives
to behave as logical AND instead of OR. Now, it is
possible to opt-out of building and using any
platform-specific asm code.

Further, this commit removes the travis CI configuration
and switches to Github Actions.
