--- .travis.yml
@@ -1,24 +1,0 @@
-go_import_path: github.com/minio/highwayhash
-sudo: required
-
-dist: trusty
-
-language: go
-
-os:
-- linux
-
-env:
-- ARCH=x86_64
-- ARCH=i686
-
-go:
-- "1.7"
-- "1.8.7"
-- "1.9.4"
-- "1.10"
-
-script:
-- diff -au <(gofmt -d .) <(printf "")
-- go vet ./...
-- go test -v ./...
