+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/minio/highwayhash
 
+go 1.15
+
 require golang.org/x/sys v0.0.0-20190130150945-aca44879d564
