+++ highwayhash_arm64.go
@@ -1,8 +1,8 @@
-//+build !noasm
-
 // Copyright (c) 2017 Minio Inc. All rights reserved.
 // Use of this source code is governed by a license that can be
 // found in the LICENSE file.
+
+//+build !noasm,!appengine
 
 package highwayhash
